Reorganizing files and filenames for readibility - this may look scary, but it's better when it is done not too late in the project.
