--- src/App.scss
@@ -1,37 +1,0 @@
-.App {
-  background-color: lightgrey;
-  display: flex;
-  flex-direction: column;
-  min-height: 100vh;
-}
-
-.App-header {
-  display: flex;
-
-  h1 {
-    margin: 32px 0 0 0;
-  }
-}
-
-.App-logo {
-  pointer-events: none;
-  width: 160px;
-}
-
-@media (prefers-reduced-motion: no-preference) {
-
-  .App-logo {
-    animation: App-logo-spin infinite 20s linear;
-  }
-}
-
-@keyframes App-logo-spin {
-
-  from {
-    transform: rotate(0deg);
-  }
-
-  to {
-    transform: rotate(360deg);
-  }
-}
