+++ src/components/App/App.scss
@@ -1,0 +1,37 @@
+.App {
+  background-color: lightgrey;
+  display: flex;
+  flex-direction: column;
+  min-height: 100vh;
+
+  &-header {
+    display: flex;
+
+    h1 {
+      margin: 32px 0 0 0;
+    }
+  }
+
+  &-logo {
+    pointer-events: none;
+    width: 160px;
+  }
+
+  @media (prefers-reduced-motion: no-preference) {
+
+    &-logo {
+      animation: App-logo-spin infinite 20s linear;
+    }
+  }
+}
+
+@keyframes App-logo-spin {
+
+  from {
+    transform: rotate(0deg);
+  }
+
+  to {
+    transform: rotate(360deg);
+  }
+}
