+++ src/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import './index.scss';
-import App from './App';
+import './assets/styles/index.scss';
+import App from './components/App/App';
 
 ReactDOM.render(
   <React.StrictMode>
